# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PICKSURE - ZIEL DOCUMENTATIONS",
        "contact": {
            "email": "info@ziel.com.co"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/auth/token": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login de Usuario y Generación de Token",
                "description": "Retorna un Json con el Token del Usuario",
                "operationId": "8f34aba8da985e9d709116a11c59d3c1",
                "requestBody": {
                    "content": {
                        "application/raw": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "flover.sanchez@ziel.com.co"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Resultado de la Operación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "msg": {
                                            "type": "string",
                                            "example": "TTpPem996u4yMZ"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Estado Invalido de la Operación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Los datos son incorrectos."
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": "..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/categories": {
            "get": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Devuelve todos los Categorias filtrando el lenguage recibido",
                "description": "Retorna un Json con los Datos de las Categorias.",
                "operationId": "9e975d8e2b6733988615c39b97de2eec",
                "parameters": [
                    {
                        "name": "lenguage",
                        "in": "query",
                        "description": "ID del Lenguage",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "default": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resultado de la Operación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "": {
                                            "type": "array",
                                            "items": {
                                                "type": "array",
                                                "items": {}
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Estado Invalido de la Operación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Los datos son incorrectos."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/categories/{prefijo}": {
            "get": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Devuelve todos los Municipios asignados a un departamento",
                "description": "Retorna un Json con los Datos de los Municipios.",
                "operationId": "0242e3b8b81f18a1ebf86ad99ce2101b",
                "parameters": [
                    {
                        "name": "prefijo",
                        "in": "path",
                        "description": "prefijo",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "ES"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resultado de la Operación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Cali"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Estado Invalido de la Operación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Los datos son incorrectos."
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": "..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/department": {
            "get": {
                "tags": [
                    "Departamento"
                ],
                "summary": "Devuelve todos los Departamentos",
                "description": "Retorna un Json con los Datos de los Departamentos.",
                "operationId": "29ef540def3a04356f8ccb0e8c44f45e",
                "responses": {
                    "200": {
                        "description": "Resultado de la Operación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Valle del Cauca"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Estado Invalido de la Operación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Los datos son incorrectos."
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": "..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/imageproducts": {
            "get": {
                "tags": [
                    "Imagenes"
                ],
                "summary": "Devuelve todas las imagenes filtrando el lenguaje",
                "description": "Retorna un Json con los titulos de las imagenes filtradas por lenguaje",
                "operationId": "0ed13cca4ae12811f15a90d3c28ddec9",
                "parameters": [
                    {
                        "name": "lenguage",
                        "in": "query",
                        "description": "ID del Lenguage",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "default": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resultado de la Operación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "title de la imagen": {
                                            "type": "varchar",
                                            "example": "String"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Estado Invalido de la Operación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Lenguaje no existe."
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": "..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/imageproducts/{id}": {
            "get": {
                "tags": [
                    "Imagenes"
                ],
                "summary": "Devuelve una imagen por su Id",
                "description": "Retorna un Json con la imagen seleccionada por el Id",
                "operationId": "0fc067d089cb61eaac401fad62d93a84",
                "parameters": [
                    {
                        "name": "Id imagen",
                        "in": "path",
                        "description": "Id imagen",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resultado de la Operación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "imagen"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Estado Invalido de la Operación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Los datos son incorrectos."
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": "..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/town/department/{idDepartment}": {
            "get": {
                "tags": [
                    "Municipios"
                ],
                "summary": "Devuelve todos los Municipios asignados a un departamento",
                "description": "Retorna un Json con los Datos de los Municipios.",
                "operationId": "index",
                "parameters": [
                    {
                        "name": "idDepartment",
                        "in": "path",
                        "description": "ID del Departamento",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "default": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resultado de la Operación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Cali"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Estado Invalido de la Operación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Los datos son incorrectos."
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": "..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token de acceso obtenido en la autenticación",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "Bearer"
            }
        }
    }
}